+++ build_modules/lib/build_modules.dart
@@ -1,0 +1,17 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+export 'src/kernel_builder.dart'
+    show KernelSummaryBuilder, kernelSummaryExtension, multiRootScheme;
+export 'src/module_builder.dart' show ModuleBuilder, moduleExtension;
+export 'src/modules.dart';
+export 'src/scratch_space.dart' show scratchSpace, scratchSpaceResource;
+export 'src/summary_builder.dart'
+    show
+        LinkedSummaryBuilder,
+        UnlinkedSummaryBuilder,
+        linkedSummaryExtension,
+        unlinkedSummaryExtension;
+export 'src/workers.dart'
+    show dart2JsWorkerResource, dartdevcDriverResource, dartdevkDriverResource;