+++ build_modules/test/fixtures/a/lib/a_cycle.dart
@@ -1,0 +1,5 @@
+// ignore_for_file: unused_import
+import 'package:b/b_cycle_with_a.dart';
+import 'a_secondary_in_cycle.dart';
+
+part 'a_part_in_cycle.dart';