+++ build_modules/test/fixtures/b/lib/b_cycle_with_a.dart
@@ -1,0 +1,2 @@
+// ignore: unused_import
+import 'package:a/a_cycle_with_b.dart';