+++ e2e_example/tool/build_kernel.dart
@@ -6,6 +6,7 @@ import 'dart:async';
 
 import 'package:build_config/build_config.dart';
 import 'package:build_web_compilers/build_web_compilers.dart';
+import 'package:build_modules/build_modules.dart';
 import 'package:build_runner/build_runner.dart';
 import 'package:build_test/builder.dart';
 
@@ -17,10 +18,17 @@ Future main(List<String> args) async {
             const InputSet(include: const ['test/**_test.dart']),
         hideOutput: true),
     apply(
-        'build_compilers|ddc',
+        'build_modules|modules',
         [
           (_) => new ModuleBuilder(),
           (_) => new KernelSummaryBuilder(),
+        ],
+        toAllPackages(),
+        isOptional: true,
+        hideOutput: true),
+    apply(
+        'build_web_compilers|ddc',
+        [
           (_) => new DevCompilerBuilder(useKernel: true),
         ],
         toAllPackages(),