+++ build_web_compilers/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3.0
+
+## Breaking changes
+
+- Split `ModuleBuilder`, `UnlinkedSummaryBuilder` and `LinkedSummaryBuilder`
+  into a separate `build_modules` package.
+
 # 0.2.1+1
 
 - Support the latest `analyzer` package.